Imported specification from internal documentation
I cleaned the specification of internal data and rewritten some
sentences to reflect the fact that this document is now published
under an open license.

Since this document is now public, old versioning will have no
meaning to the rest of the world, so I took the liberty to reset
version to 0.1 with a draft status.